requirements.txt
requests
python-dotenv
supabase
anthropic
pytest
